src/requirements.txt
pyyaml
rank_bm25
scikit-learn
rpyc
redis
sentencepiece
rpyc
TexSoup==0.1.4
psycopg2-binary
sqlalchemy
pudb
pandas
pyahocorasick
nltk
tex2py
requests
sumy
numpy
